backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1,<4.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
djangorestframework-simplejwt>=5.3,<6.0
dj-database-url>=1.3,<2.0
python-dotenv>=1.0,<2.0
argon2-cffi>=23.1,<24.0
Pillow>=10.0,<11.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
